Add ApplyFilters extension method for dynamic query filtering; replace GetAllPagenatedHandler with GetBranchesPagenatedHandler
